requirements.txt
discord.py
jishaku
feedparser
